requirements.txt
python-dotenv==0.9.1
requests==2.27.1